physical-ai-textbook/rag-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
qdrant-client==1.7.0
openai==1.10.0
python-dotenv==1.0.0
pydantic==2.5.3
google-generativeai==0.3.2
requests==2.31.0
